move things around. Fixes from rebase
