+++ royals_core/src/game_state.rs
@@ -4,13 +4,12 @@ use strum::IntoEnumIterator;
 
 use crate::{
     card::Card,
-    event::{Event, EventEntry, EventVisibility},
-    play::{Action, ActionId, Play},
+    event::EventEntry,
+    play::{Action, Play},
     player::PlayerId,
-    utils::VecExtensions,
 };
 
-struct PlayerState {
+pub struct PlayerState {
     protected: bool,
     hand: Vec<Card>,
 }
@@ -22,32 +21,32 @@ impl PlayerState {
             hand: vec![],
         }
     }
-    fn protected(&self) -> bool {
-        self.protected.clone()
+    pub fn protected(&self) -> bool {
+        self.protected
     }
 
-    fn set_protected(&mut self, value: bool) {
+    pub fn set_protected(&mut self, value: bool) {
         self.protected = value;
     }
 
-    fn hand(&self) -> &Vec<Card> {
+    pub fn hand(&self) -> &Vec<Card> {
         &self.hand
     }
 
-    fn hand_mut(&mut self) -> &mut Vec<Card> {
+    pub fn hand_mut(&mut self) -> &mut Vec<Card> {
         &mut self.hand
     }
 
-    fn is_active(&self) -> bool {
+    pub fn is_active(&self) -> bool {
         !&self.hand().is_empty()
     }
 }
 
 pub struct GameState<'a> {
-    players: Vec<PlayerState>,
-    deck_head: usize,
-    players_turn: PlayerId,
-    deck: &'a [Card],
+    pub players: Vec<PlayerState>,
+    pub deck_head: usize,
+    pub players_turn: PlayerId,
+    pub deck: &'a [Card],
 }
 
 impl<'a> GameState<'a> {
@@ -65,22 +64,6 @@ impl<'a> GameState<'a> {
         state.pick_up_card(state.players_turn, log);
         state
     }
-
-    pub fn handle_action(&mut self, action: ActionId, log: &mut Vec<EventEntry>) {
-        let (_, actions) = self.valid_actions();
-        if action < actions.len() {
-            match &actions[action] {
-                Action::GiveUp => {
-                    self.drop_player(self.players_turn, "Player gave up", log);
-                }
-                Action::Play(p) => {
-                    self.handle_play(p, log);
-                }
-            }
-            self.next_player_turn(log);
-        }
-    }
-
     pub fn valid_actions(&self) -> (Option<PlayerId>, Vec<Action>) {
         let actions: Vec<Action> = once(Action::GiveUp)
             .chain(self.possible_actions())
@@ -96,13 +79,13 @@ impl<'a> GameState<'a> {
         )
     }
 
-    fn possible_actions(&self) -> Vec<Action> {
+    pub fn possible_actions(&self) -> Vec<Action> {
         // todo is there an alternative way to also iterate over None
         let others = self.other_active_players();
-        let mut optional_players = others.iter().map(|p| Some(p.clone())).collect_vec();
+        let mut optional_players = others.iter().map(|p| Some(*p)).collect_vec();
         optional_players.push(None);
 
-        let mut optional_card = Card::iter().map(|c| Some(c)).collect_vec();
+        let mut optional_card = Card::iter().map(Some).collect_vec();
         optional_card.push(None);
 
         iproduct!(Card::iter(), optional_players.iter(), optional_card.iter())
@@ -116,7 +99,7 @@ impl<'a> GameState<'a> {
             .collect_vec()
     }
 
-    fn is_valid(&self, action: &Action) -> bool {
+    pub fn is_valid(&self, action: &Action) -> bool {
         if self.game_over() {
             return false;
         }
@@ -129,29 +112,24 @@ impl<'a> GameState<'a> {
                 if self.players[self.players_turn]
                     .hand()
                     .contains(&Card::Countess)
+                    && (play.card == Card::Prince || play.card == Card::King)
                 {
-                    if play.card == Card::Prince || play.card == Card::King {
-                        return false;
-                    }
+                    return false;
                 }
                 if !play.card.needs_opponent() {
                     if play.opponent.is_some() {
                         return false;
                     }
-                } else if !self.all_protected() {
-                    if play.opponent.is_none() {
-                        return false;
-                    }
+                } else if !self.all_protected() && play.opponent.is_none() {
+                    return false;
                 }
 
                 if !play.card.needs_guess() {
                     if play.guess.is_some() {
                         return false;
                     }
-                } else if !self.all_protected() {
-                    if play.guess.is_none() {
-                        return false;
-                    }
+                } else if !self.all_protected() && play.guess.is_none() {
+                    return false;
                 }
 
                 if let Some(op) = play.opponent {
@@ -170,7 +148,7 @@ impl<'a> GameState<'a> {
         }
     }
 
-    fn active_players(&self) -> HashSet<PlayerId> {
+    pub fn active_players(&self) -> HashSet<PlayerId> {
         self.players
             .iter()
             .enumerate()
@@ -179,7 +157,7 @@ impl<'a> GameState<'a> {
             .collect()
     }
 
-    fn other_players(&self) -> HashSet<PlayerId> {
+    pub fn other_players(&self) -> HashSet<PlayerId> {
         self.players
             .iter()
             .enumerate()
@@ -188,178 +166,22 @@ impl<'a> GameState<'a> {
             .collect()
     }
 
-    fn other_active_players(&self) -> HashSet<PlayerId> {
+    pub fn other_active_players(&self) -> HashSet<PlayerId> {
         self.other_players()
             .intersection(&self.active_players())
             .cloned()
             .collect::<HashSet<_>>()
     }
 
-    fn all_protected(&self) -> bool {
+    pub fn all_protected(&self) -> bool {
         self.other_active_players()
             .iter()
             .all(|&id| self.players[id].protected())
     }
 
-    fn game_over(&self) -> bool {
+    pub fn game_over(&self) -> bool {
         // last card is ussually not used
         self.deck.len() - self.deck_head <= 1 || self.active_players().len() <= 1
-    }
-
-    fn pick_up_card(&mut self, player_id: PlayerId, log: &mut Vec<EventEntry>) {
-        let next_card = self.deck[self.deck_head];
-        self.deck_head += 1;
-        log.push(EventEntry {
-            visibility: EventVisibility::Private(player_id),
-            event: Event::PickUp(
-                player_id,
-                Some(next_card.clone()),
-                self.deck.len() - self.deck_head,
-            ),
-        });
-        self.players[player_id].hand_mut().push(next_card);
-    }
-
-    fn drop_player(&mut self, player_id: PlayerId, reason: &str, log: &mut Vec<EventEntry>) {
-        while let Some(op_card) = self.players[player_id].hand_mut().pop() {
-            log.push(EventEntry {
-                visibility: EventVisibility::Public,
-                event: Event::Fold(player_id, op_card, reason.to_string()),
-            });
-        }
-        log.push(EventEntry {
-            visibility: EventVisibility::Public,
-            event: Event::DropOut(player_id),
-        });
-    }
-
-    fn next_player_turn(&mut self, log: &mut Vec<EventEntry>) {
-        self.players_turn = (self.players_turn + 1) % self.players.len();
-        while !self.players[self.players_turn].is_active() {
-            self.players_turn = (self.players_turn + 1) % self.players.len();
-        }
-        if !self.game_over() {
-            self.pick_up_card(self.players_turn, log);
-        } else {
-            self.wrap_up_round(log);
-        }
-    }
-
-    fn wrap_up_round(&mut self, log: &mut Vec<EventEntry>) {
-        let mut best_players: Vec<PlayerId> = vec![];
-        let mut best_card: Option<Card> = None;
-        for (i, p) in self.players.iter().enumerate() {
-            if let Some(player_card) = p.hand().get(0) {
-                log.push(EventEntry {
-                    visibility: EventVisibility::Public,
-                    event: Event::Fold(i, player_card.clone(), "game is finished".to_string()),
-                });
-                if let Some(card) = best_card {
-                    if card < *player_card {
-                        best_players = vec![i];
-                        best_card = Some(player_card.clone());
-                    } else if card == *player_card {
-                        best_players.push(i);
-                    }
-                } else {
-                    best_players = vec![i];
-                    best_card = Some(player_card.clone());
-                }
-            }
-        }
-
-        log.push(EventEntry {
-            visibility: EventVisibility::Public,
-            event: Event::Winner(best_players),
-        });
-
-        for e in log {
-            e.visibility = EventVisibility::Public;
-        }
-    }
-
-    fn handle_play(&mut self, p: &Play, log: &mut Vec<EventEntry>) {
-        let card = self.players[self.players_turn]
-            .hand_mut()
-            .remove_first_where(|&card| card == p.card)
-            .unwrap();
-
-        log.push(EventEntry {
-            visibility: EventVisibility::Public,
-            event: Event::Play(self.players_turn, p.clone()),
-        });
-        if let Some(opponent) = p.opponent {
-            // do not attack protected player
-            if self.players[opponent].protected() && !self.all_protected() {
-                self.drop_player(self.players_turn, "attacked a protected player", log);
-                return;
-            }
-        }
-        self.players[self.players_turn].set_protected(false);
-        match card {
-            Card::Guard => {
-                if let Some(op) = p.opponent {
-                    let g = p.guess.unwrap();
-                    if self.players[op].hand()[0] == g {
-                        self.drop_player(op, "opponent guessed the hand card", log)
-                    }
-                }
-            }
-            Card::Priest => {
-                if let Some(op) = p.opponent {
-                    log.push(EventEntry {
-                        visibility: EventVisibility::Private(self.players_turn),
-                        event: Event::LearnedCard(op, Some(self.players[op].hand()[0].clone())),
-                    });
-                }
-            }
-            Card::Baron => {
-                if let Some(op) = p.opponent {
-                    let op_card = self.players[op].hand()[0];
-                    let player_card = self.players[self.players_turn].hand()[0];
-                    if op_card < player_card {
-                        self.drop_player(op, "smaller card then opponent", log);
-                    } else if player_card < op_card {
-                        self.drop_player(self.players_turn, "smaller card then opponent", log);
-                    }
-                }
-            }
-            Card::Maid => {
-                self.players[self.players_turn].set_protected(true);
-            }
-            Card::Prince => {
-                if let Some(op) = p.opponent {
-                    if self.players[op].hand()[0] == Card::Princess {
-                        self.drop_player(op, "forced to play the princess", log);
-                    } else {
-                        let folded = self.players[op].hand_mut().pop().unwrap();
-                        log.push(EventEntry {
-                            visibility: EventVisibility::Public,
-                            event: Event::Fold(
-                                op,
-                                folded,
-                                "opponent has played prince to force it".to_string(),
-                            ),
-                        });
-                        self.pick_up_card(op, log);
-                    }
-                }
-            }
-            Card::King => {
-                if let Some(op) = p.opponent {
-                    let op_card = self.players[op].hand_mut().pop().unwrap();
-                    let player_card = self.players[self.players_turn].hand_mut().pop().unwrap();
-                    self.players[op].hand_mut().push(player_card);
-                    self.players[self.players_turn].hand_mut().push(op_card);
-                }
-            }
-            Card::Countess => {}
-            Card::Princess => self.drop_player(
-                self.players_turn,
-                "playing the princess is equivalent to giving up",
-                log,
-            ),
-        }
     }
 }
